Move HTML data into a separate module
